Separate the business logic from the view of the transaction.
